% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_chr_words.R
\name{count_chr_words}
\alias{count_chr_words}
\title{Counts words and characters in a string}
\usage{
count_chr_words(text, count_spaces)
}
\arguments{
\item{text}{The string whose words and characters will be counted.
Parameter is called text since it makes intuitive sense to the user (they are inputting written text).}

\item{count_spaces}{Weather spaces in the string should be counted as characters \strong{(TRUE or FALSE)}.
count_spaces is pretty self explanatory. We are asking the user if they
should count spaces as character or not.}
}
\value{
A numeric vector whose first element
is the number of words in the string and second element is number
of characters in the string. \code{c(words, characters)}
}
\description{
This function counts the number of words
and characters in a given string. The user can decide
weather to count spaces as characters or not.
}
\examples{
count_chr_words("hello world", count_spaces = TRUE)
count_chr_words("AATTCCTTTCGGTG", count_spaces = FALSE)

}
